pyproject.toml
[project]
name = "image-resizer"
version = "0.1.0"
description = "Image resizing and optimization script"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Pillow",
    "requests",
    "python-dotenv"
]
